Fix multiple division-by-zero conditions in MPEG4 parsing
Several situations arise processing MP4 atoms that lead to undefined behavior
when dividing by zero. Typically this results in a crash (denial of service
condition).

NOTE: In most cases we simply avoid the division, leaving kKeyDuration unset.
It may be more desirable to bail out, as we do in the parseSegmentIndex case.

Change-Id: Iab0118934b49eabd0b44be8408160d7f122c28df

Signed-off-by: zephiK <zephik@gmail.com>